+++ patches/src/main/kotlin/app/revanced/patches/samsung/radio/misc/fix/crash/Fingerprints.kt
@@ -1,0 +1,12 @@
+package app.revanced.patches.samsung.radio.misc.fix.crash
+
+import app.revanced.patcher.fingerprint
+
+internal val permissionRequestListFingerprint by fingerprint {
+    strings(
+        "android.permission.POST_NOTIFICATIONS",
+        "android.permission.READ_MEDIA_AUDIO",
+        "android.permission.RECORD_AUDIO"
+    )
+    custom { method, _ ->  method.name == "<clinit>" }
+}